requirements.txt
mysql-connector-python==8.1.0
pillow==10.0.0
